StoreTimeSpansAsNumbers with a factory.
